refactor social media linking to redux thunk
